requirements.txt
# ============================================================
# battery_ts_rag — numerical battery SOH forecasting framework
# ============================================================
# Requires Python >= 3.10

# --- Vector search ------------------------------------------
faiss-cpu>=1.7.2

# --- Numerical / data ---------------------------------------
numpy>=1.25.0
pandas>=2.0.0
pyarrow>=12.0.0          # for parquet I/O

# --- Deep learning ------------------------------------------
torch>=2.0.0

# --- Config / CLI -------------------------------------------
pyyaml>=6.0

# --- Progress bars ------------------------------------------
tqdm>=4.60.0

# --- Plotting -----------------------------------------------
matplotlib>=3.8.0

# --- (Optional) experiment tracking -------------------------
# wandb>=0.19.0
